refactor(#319): split state-utils.ts into focused modules
Split the 943-line state-utils.ts into focused, single-responsibility modules:

- pr-status.ts (151 lines): PR merge detection and branch status
- state-rebuild.ts (194 lines): State reconstruction from run logs
- worktree-discovery.ts (320 lines): Worktree discovery for state bootstrapping
- state-cleanup.ts (331 lines): Cleanup of stale entries and startup reconciliation
- state-utils.ts (53 lines): Re-export barrel for backward compatibility

All existing imports continue to work through the barrel file.

Entire-Checkpoint: be89c6723d82
